fix direnv execution on precm_functions
It was too slow. It is now faster: as I installed direnv from nix and removed it from brew, the direnv command that was picked was the one installed by asdf. It has ots of indirections (shims) and was taking +150ms to run. Using 'direnv hook zsh' as already done in direnv/load.zsh sets up a hardcoded path for direnv which is was way faster than going through all indirection layers of asdf.